refactoring due to new general file with constants
