+++ controllers/category/addCategoryController.js
@@ -1,7 +1,6 @@
 const { Category } = require('../../models')
 
 const addCategory = async (req, res) => {
-  console.log(req.body)
   const result = await Category.create(req.body)
 
   const { _id, type, name, iconUrl } = result
