+++ middlewares/validation/index.js
@@ -1,10 +1,8 @@
-const ERROR_MESSAGES = require('./validationErrorMessages')
 const transactionsValidation = require('./transactions')
 const reportsValidation = require('./reports')
 const categoryValidation = require('./categories')
 
 module.exports = {
-  ERROR_MESSAGES,
   transactionsValidation,
   reportsValidation,
   categoryValidation
