+++ middlewares/validation/reports/summaryReportGetting/querySchema.js
@@ -1,11 +1,6 @@
 const Joi = require('joi')
 
-const ERROR_MESSAGES = require('../../validationErrorMessages')
-
-const TRANSACTION_TYPES = {
-  DEBIT: 'debit',
-  CREDIT: 'credit'
-}
+const { TRANSACTION_TYPES, ERROR_MESSAGES } = require('../../../../utils')
 
 const querySchema = Joi.object({
   type: Joi.string()
@@ -23,18 +18,6 @@ const querySchema = Joi.object({
       'any.required': ERROR_MESSAGES.missingDateField,
       'date.format': ERROR_MESSAGES.invalidDateFormat
     })
-
-  // Joi.number()
-  // .required()
-  // .integer()
-  // .min(1900)
-  // .max(9999)
-  // .messages({
-  //   'any.required': ERROR_MESSAGES.missingPeriodQueryString,
-  //   'number.base': ERROR_MESSAGES.invalidPeriodQueryStringFormat,
-  //   'number.min': ERROR_MESSAGES.outOfRangePeriodQueryString,
-  //   'number.max': ERROR_MESSAGES.outOfRangePeriodQueryString
-  // })
 })
 
 module.exports = querySchema
