+++ utils/generalIndicatorsCounting.js
@@ -1,13 +1,10 @@
-const CATEGORY_TYPES = {
-  INCOME: 'income',
-  COST: 'cost'
-}
+const { REPORTS_CATEGORY_TYPES } = require('./constants')
 
 const generalIndicatorsCounting = report => {
   return report.reduce((acc, elem) => {
     const tmp = [...acc]
 
-    elem.type === CATEGORY_TYPES.INCOME
+    elem.type === REPORTS_CATEGORY_TYPES.INCOME
       ? tmp[0] += elem.total
       : tmp[1] += elem.total
