+++ utils/index.js
@@ -1,3 +1,4 @@
+const constants = require('./constants')
 const createUrl = require('./createUrl')
 const makeSummary = require('./makeSummary')
 const getDaysInterval = require('./getDaysInterval')
@@ -8,6 +9,7 @@ const groupByUniqueDescriptions = require('./groupByUniqueDescriptions')
 const generalIndicatorsCounting = require('./generalIndicatorsCounting')
 
 module.exports = {
+  ...constants,
   createUrl,
   makeSummary,
   getDaysInterval,
